Migrate frameSrc to TypeScript
