--- src/Playroom/Frames/frameSrc.d.ts
@@ -1,11 +1,0 @@
-interface FrameParams {
-  code: string;
-  themeName: string;
-}
-
-declare function frameSrc(
-  frameParams: FrameParams,
-  config: InternalPlayroomConfig
-): string;
-
-export default frameSrc;
